backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.2
psycopg2-binary==2.9.9
pytest==8.2.2
PyJWT==2.8.0
python-multipart==0.0.9
